+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - (spec) Added a file export resource type as a resource which packages can provide as part of their deployments and/or feature flags.
 - (cli) Added checking of conflicts between file export resources with `plt check`/`stage check`.
-- (cli) When information about a package deployment is shown (e.g. with `stage show-bun-depl` or `[dev] plt show-depl`), information about the target paths of file exports is also shown.
+- (cli) When information about a package is shown (e.g. with `cache show-pkg` or `[dev] plt show-pkg`), the source and target paths of file exports are also shown.
+- (cli) When information about a package deployment is shown (e.g. with `stage show-bun-depl` or `[dev] plt show-depl`), the target paths of file exports are also shown.
+- (cli) When information about a staged pallet bundle is shown (e.g. with `stage show-bun`), the target paths of file exports are also shown for each package deployment.
+- (cli) Information about the target paths of file exports for each package deployment in a staged pallet bundle is now recorded in the staged pallet bundle's manifest file.
 
 ## 0.7.0-alpha.0 - 2024-04-10